+++ contracts/test/unit/merkle/MerkleProof.sol
@@ -1,0 +1,41 @@
+// Copyright 2023-2024 LightDotSo.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//   http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+// SPDX-License-Identifier: Apache-2.0
+
+pragma solidity ^0.8.18;
+
+import {MerkleProof} from "@openzeppelin/contracts/utils/cryptography/MerkleProof.sol";
+import {BaseTest} from "@/test/base/BaseTest.t.sol";
+// solhint-disable-next-line no-console
+import {console} from "forge-std/console.sol";
+
+/// @notice Unit tests for `MerkleProof` for getting base hash
+contract MerkleProofUnitTest is BaseTest {
+    // -------------------------------------------------------------------------
+    // Tests
+    // -------------------------------------------------------------------------
+
+    /// Tests raw `Create2.safeCreate2`
+    function test_simple_merkle_proof() public {
+        // solhint-disable-next-line no-console
+        console.logAddress(address(this));
+
+        address computedAddress = Create2.computeAddress(bytes32(0), bytes32(0));
+
+        // Log the byte code hash
+        // solhint-disable-next-line no-console
+        console.logAddress(computedAddress);
+    }
+}
