+++ crates/contracts/src/merkle_tree.rs
@@ -79,11 +79,8 @@ impl StandardMerkleTree {
             .map(|(i, value)| (value, i, standard_leaf_hash(value)))
             .collect();
 
-        let mut hashed_values_hash =
+        let hashed_values_hash =
             hashed_values.iter().map(|(_, _, hash)| *hash).collect::<Vec<B256>>();
-
-        // Sort the hashed values
-        hashed_values_hash.sort();
 
         let tree = make_merkle_tree(hashed_values_hash);
 
@@ -275,5 +272,12 @@ mod test {
             let bool = tree.verify_proof(&leaf, proof);
             assert!(bool);
         }
+
+        let root = tree.root();
+        assert_eq!(
+            root,
+            B256::from_hex("0x0030ce873e657283a8e03a3e83ba95a0bf1ad049e8ac1cb8148280aca2b1adc7")
+                .unwrap()
+        );
     }
 }
